requirements.txt
pycryptodome
Image
opencv-python
numpy